etl/pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = ""
authors = [
    {name = "Felipe Sebben",email = "felipesebben@yahoo.com.br"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pantab (>=5.2.2,<6.0.0)",
    "tableauhyperapi (>=0.0.24081,<0.0.24082)",
    "tableauserverclient (>=0.38,<0.39)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
